If the get-queries endpoint always returns 401, the client goes into a tight loop.
The client has code in it that tries to do this every 5s, but it's
not getting triggered because this kind of auth error comes after
connect. So from client's POV, the connection is successful then
it fails with an auth error soon after.

We do plan to rewrite this code so I'm not sure what the right
thing is. I'd like to find a quick fix that is consistent with
current intent of this system - this kind of auth failure should
cause client to retry once immediately and then every 5s after.

One way to do this might be to adjust the notion of 'connected'
to not be until any initial queries are completed. Then the
existing 5s logic should work again.